+++ db/migrate/20180731155511_create_cities.rb
@@ -3,9 +3,9 @@
 class CreateCities < ActiveRecord::Migration[5.2]
   def self.up
     create_table :TB_CIDADE, id: false do |t|
-      t.integer :CO_CODIGO, null: false
+      t.bigint :CO_CODIGO, null: false
       t.string :NO_NOME, null: false
-      t.integer :CO_UF, null: false
+      t.bigint :CO_UF, null: false
     end
 
     add_foreign_key :TB_CIDADE, :TB_UF, column: :CO_UF, primary_key: :CO_CODIGO
