+++ db/migrate/20181105181800_edit_attachment_links_with_feedbacks.rb
@@ -2,7 +2,7 @@
 
 class EditAttachmentLinksWithFeedbacks < ActiveRecord::Migration[5.2]
   def change
-    add_column :RT_LINK_ANEXO, :CO_FEEDBACK, :integer
+    add_column :RT_LINK_ANEXO, :CO_FEEDBACK, :bigint
     add_foreign_key :RT_LINK_ANEXO, :TB_FEEDBACK, column: :CO_FEEDBACK, primary_key: :CO_SEQ_ID
   end
 end
